MethodsDebuggingAndTroubleshootingCode final solutions

+++ MethodsDebuggingAndTroubleshootingCode/InstructionSet/16.StartUp.cs
@@ -5,6 +5,7 @@ class InstructionSet_fixed
     static void Main()
     {
         string opCode = Console.ReadLine();
+        //string opCode = "INC " + int.MaxValue;
 
         while (opCode != "END")
         {
@@ -15,20 +16,20 @@ class InstructionSet_fixed
             {
                 case "INC":
                     {
-                        int operandOne = int.Parse(codeArgs[1]);
+                        long operandOne = long.Parse(codeArgs[1]);
                         result = ++operandOne;
                         break;
                     }
                 case "DEC":
                     {
-                        int operandOne = int.Parse(codeArgs[1]);
+                        long operandOne = long.Parse(codeArgs[1]);
                         result = --operandOne;
                         break;
                     }
                 case "ADD":
                     {
-                        int operandOne = int.Parse(codeArgs[1]);
-                        int operandTwo = int.Parse(codeArgs[2]);
+                        long operandOne = long.Parse(codeArgs[1]);
+                        long operandTwo = long.Parse(codeArgs[2]);
                         result = operandOne + operandTwo;
                         break;
                     }
